fix: move `setDashboardFromFile` function from `libs` to `libs/game`
